+++ src/skiplist.h
@@ -15,7 +15,7 @@ extern skiplist sl_create2(int (*compare)(const void*, const void*), void* (*cop
 extern int sl_add(skiplist s, const void* key, const void* value);
 extern int sl_rem(skiplist s, const void* key);
 extern int sl_get(const skiplist s, const void* key, const void** value);
-extern size_t sl_count(const skiplist s);
+extern unsigned long sl_count(const skiplist s);
 
 // Iterate over the whole range. f returns zero to halt.
