+++ src/tree.h
@@ -1,8 +1,6 @@
 #ifndef TREE_H
 #define TREE_H
 
-#include <stdlib.h>
-
 #ifndef TREE_KEY_USERDEFINED
 #include "uuid.h"
 #define TREE_KEY uuid
@@ -18,9 +16,9 @@ extern int tree_get(const tree tptr, const TREE_KEY* key, unsigned long long* va
 extern int tree_set(const tree tptr, const TREE_KEY* key, unsigned long long value);
 extern int tree_del(tree tptr, const TREE_KEY* key);
 
-extern size_t tree_count(const tree tptr);
-extern int tree_stats(const tree tptr, long* trunks, long* branches, long* leafs);
-extern size_t tree_iter(const tree tptr, void* h, int (*f)(void* h, const TREE_KEY* key, unsigned long long* value));
+extern unsigned long tree_count(const tree tptr);
+extern int tree_stats(const tree tptr, unsigned long* trunks, unsigned long* branches, unsigned long* leafs);
+extern unsigned long tree_iter(const tree tptr, void* h, int (*f)(void* h, const TREE_KEY* key, unsigned long long* value));
 
 extern void tree_destroy(tree tptr);
